Adding the two TI Cartridge case files (STL)s
This is a top and bottom STL for the TI cartridge shell sized for standard TI-99/4A cartridges. Minor adjustments made at the 2014 TI-99/4A Faire in Chicago.  This is the last version that I authored of this shell.  Measured with a micrometer and edited to add screw holes instead of the clasps to hold it shut.

Notes:
This is in inches; not metric.  When we first started doing 3D modeling like 20 years ago, the instructor (mistakenly) had us do it in inches.  If someone can please help me to convert this to metric, and help me fix the rough top, I'd be happy to accept a merge request to fix it.  (I don’t have a license to Inventor any longer, so I cannot modify it.)

Knock yourself out and do whatever you want.  That’s why it’s here!  License: CC BY-SA